Pulled currencyPair to operations folder
