+++ src/client/src/combineReducers.ts
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux'
 import compositeStatusServiceReducer from './compositeStatusServiceOperations'
 import { connectionStatusReducer } from './connectionStatusOperations'
-import { currencyPairReducer } from './currencyPairsOperations'
 import { enviromentReducer } from './enviromentReducer'
+import { currencyPairReducer } from './operations/currencyPairs'
 import { pricingServiceReducer } from './operations/pricing'
 import { analyticsReducer } from './ui/analytics'
 import { blotterReducer } from './ui/blotter'
