+++ src/client/src/operations/currencyPairs/index.ts
@@ -1,0 +1,1 @@
+export { currencyPairReducer, CurrencyPairState } from './reducer'
