+++ src/client/src/operations/currencyPairs/utils.ts
@@ -1,0 +1,8 @@
+// TODO: revisit
+export function toObject<T>(aMap: Map<string, T>): { [id: string]: T } {
+  const obj = {}
+  aMap.forEach((v, k) => {
+    obj[k] = v
+  })
+  return obj
+}
