+++ src/client/src/referenceDataOperations.ts
@@ -1,6 +1,6 @@
-import { createAction } from 'redux-actions'
 import { ofType } from 'redux-observable'
 import { map, switchMapTo, takeUntil } from 'rxjs/operators'
+import { action } from './ActionHelper'
 import { ApplicationEpic } from './ApplicationEpic'
 import { ACTION_TYPES as CONNECTION_ACTION_TYPES } from './connectionActions'
 import { CurrencyPair } from './types'
@@ -9,7 +9,11 @@ export enum ACTION_TYPES {
   REFERENCE_SERVICE = '@ReactiveTraderCloud/REFERENCE_SERVICE'
 }
 
-export const createReferenceServiceAction = createAction<Map<string, CurrencyPair>>(ACTION_TYPES.REFERENCE_SERVICE)
+export const createReferenceServiceAction = action<typeof ACTION_TYPES.REFERENCE_SERVICE, Map<string, CurrencyPair>>(
+  ACTION_TYPES.REFERENCE_SERVICE
+)
+
+export type ReferenceServiceAction = ReturnType<typeof createReferenceServiceAction>
 
 export const referenceServiceEpic: ApplicationEpic = (action$, store, { referenceDataService }) => {
   return action$.pipe(
